requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
branca==0.4.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
docutils==0.16
Flask==1.1.2
folium==0.11.0
future==0.18.2
Geohash==1.0
geojson==2.5.0
google-api-core==1.22.0
google-auth==1.20.1
google-cloud-bigquery==1.26.1
google-cloud-bigquery-storage==1.0.0
google-cloud-core==1.4.1
google-crc32c==0.1.0
google-resumable-media==0.7.1
googleapis-common-protos==1.52.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
MarkupSafe==1.1.1
numpy==1.19.1
pandas==1.1.0
patsy==0.5.1
Pillow==7.2.0

pytz==2020.1
requests==2.24.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.2
Shapely==1.7.0
six==1.15.0
sklearn==0.0
statsmodels==0.11.1
threadpoolctl==2.1.0
torch==1.6.0+cpu
torchvision==0.7.0+cpu
urllib3==1.25.10
Werkzeug==1.0.1
xlrd==1.2.0
#polygon-geohasher==0.0.1
protobuf==3.12.4
pyarrow==1.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
#python-geohash==0.8.5